Fixed undeclared identifier issues with pool code and added pool test function to main.
