+++ ChildrenOfOsi/enc_temp_folder/40c6cd1a55b35339b7bd4a69c3429ed/Main.cpp
